app/kubestack/requirements.txt
wheel==0.23.0
gear
python-jenkins
pyyaml
PrettyTable>=0.6,<0.8
zmq
